@@ -13,7 +13,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-from typing import Dict, Any, List, Optional
+from typing import Any, Optional
 
 from splunk_add_on_ucc_framework.entity.entity import Entity
 from splunk_add_on_ucc_framework.entity.interval_entity import IntervalEntity
@@ -26,8 +26,8 @@ ENTITY_TYPES = [
 
 
 def resolve_entity(
-    entity_definition: Dict[str, Any],
-    entity_types: Optional[List[Any]] = None,
+    entity_definition: dict[str, Any],
+    entity_types: Optional[list[Any]] = None,
 ) -> Entity:
     entity_types = entity_types if entity_types else ENTITY_TYPES
     for entity_type in entity_types:
@@ -39,12 +39,12 @@ def resolve_entity(
     return Entity(entity_definition)
 
 
-def expand_entity(entity_definition: Dict[str, Any]) -> Dict[str, Any]:
+def expand_entity(entity_definition: dict[str, Any]) -> dict[str, Any]:
     return resolve_entity(entity_definition).long_form()
 
 
 def collapse_entity(
-    entity_definition: Dict[str, Any],
-    entity_types: Optional[List[Any]] = None,
-) -> Dict[str, Any]:
+    entity_definition: dict[str, Any],
+    entity_types: Optional[list[Any]] = None,
+) -> dict[str, Any]:
     return resolve_entity(entity_definition, entity_types).short_form()
