@@ -13,13 +13,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-from typing import Optional, Dict, Any
+from typing import Optional, Any
 
 from splunk_add_on_ucc_framework.entity.entity import Entity
 
 
 class IndexEntity(Entity):
-    def long_form(self) -> Dict[str, Any]:
+    def long_form(self) -> dict[str, Any]:
         definition = {
             "type": "singleSelect",
             "field": self["field"],
@@ -53,7 +53,7 @@ class IndexEntity(Entity):
         return definition
 
     @classmethod
-    def from_definition(cls, definition: Dict[str, Any]) -> Optional[Entity]:
+    def from_definition(cls, definition: dict[str, Any]) -> Optional[Entity]:
         if definition.get("type") == "index":
             return cls(definition)
 
