@@ -13,7 +13,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 #
-from typing import Dict, List, Optional
+from typing import Optional
 
 from splunk_add_on_ucc_framework.generators.file_generator import FileGenerator
 from splunk_add_on_ucc_framework.global_config import GlobalConfig
@@ -34,7 +34,7 @@ class CommandsConf(FileGenerator):
             for command in global_config.custom_search_commands:
                 self.command_names.append(command["commandName"])
 
-    def generate(self) -> Optional[List[Dict[str, str]]]:
+    def generate(self) -> Optional[list[dict[str, str]]]:
         if not self._global_config.has_custom_search_commands():
             return None
 
