@@ -18,7 +18,7 @@ from splunk_add_on_ucc_framework.commands.modular_alert_builder import (
     arf_consts as ac,
     normalize,
 )
-from typing import Dict, Any, List, Optional
+from typing import Any, Optional
 from os import linesep
 from re import search
 from splunk_add_on_ucc_framework.global_config import GlobalConfig
@@ -43,11 +43,11 @@ class AlertActionsHtml(FileGenerator):
             schema_content = envs["schema.content"]
             self._alert_settings = schema_content["modular_alerts"]
 
-    def generate(self) -> Optional[List[Dict[str, str]]]:
+    def generate(self) -> Optional[list[dict[str, str]]]:
         if not self._global_config.has_alerts():
             return None
 
-        alert_details: List[Any] = []
+        alert_details: list[Any] = []
         for self.alert in self._alert_settings:
             self.set_template_and_render(
                 template_file_path=["html_templates"],
