@@ -14,7 +14,7 @@
 # limitations under the License.
 #
 from splunk_add_on_ucc_framework.generators.file_generator import FileGenerator
-from typing import Dict, List, Optional
+from typing import Optional
 import os
 from xml.etree.ElementTree import Element, SubElement, tostring
 from splunk_add_on_ucc_framework.utils import pretty_print_xml
@@ -106,7 +106,7 @@ class DefaultXml(FileGenerator):
         nav_as_string = tostring(nav, encoding="unicode")
         return pretty_print_xml(nav_as_string)
 
-    def generate(self) -> Optional[List[Dict[str, str]]]:
+    def generate(self) -> Optional[list[dict[str, str]]]:
         if not self._global_config.has_pages():
             return None
         file_path = self.get_file_output_path(
