@@ -14,7 +14,7 @@
 # limitations under the License.
 #
 from splunk_add_on_ucc_framework.generators.file_generator import FileGenerator
-from typing import Dict, List, Optional
+from typing import Optional
 from xml.etree.ElementTree import Element, SubElement, tostring
 from splunk_add_on_ucc_framework.utils import pretty_print_xml
 from splunk_add_on_ucc_framework.global_config import GlobalConfig
@@ -53,7 +53,7 @@ class RedirectXml(FileGenerator):
         view_as_string = tostring(view, encoding="unicode")
         return pretty_print_xml(view_as_string)
 
-    def generate(self) -> Optional[List[Dict[str, str]]]:
+    def generate(self) -> Optional[list[dict[str, str]]]:
         if not self._global_config.has_oauth():
             return None
         file_name = f"{self.ta_name}_redirect.xml"
