@@ -16,7 +16,8 @@
 import argparse
 import re
 import sys
-from typing import Optional, Sequence
+from typing import Optional
+from collections.abc import Sequence
 import logging
 
 from splunk_add_on_ucc_framework.commands import build
