@@ -1,7 +1,8 @@
 import itertools
 import json
 import os.path
-from typing import Any, Iterator
+from typing import Any
+from collections.abc import Iterator
 
 import pytest
 from unittest import mock
